fix: 修正response欄位型別錯誤

+++ backend2/src/main/java/com/cmoney/backend2/forumocean/service/api/rank/getcommodityrank/GetCommodityRankResponseBody.kt
@@ -14,7 +14,7 @@ data class GetCommodityRankResponseBody(
     @SerializedName("commodityKey")
     val commodityKey :String?,
     @SerializedName("score")
-    val score : Int?,
+    val score : Long?,
     @SerializedName("ranking")
     val ranking :Int?,
     @SerializedName("lastRanking")
